Extracted some common code.
Sometimes it would be nice to know context during IR rewriting, this is when
IrElementTransformerVoidWithContext comes in handy.
